Revert "Fix WGC display selection and webcam sync"
This reverts commit e5becc30a3bd3f1dab2f165f150ecec1d70aec3f.

+++ electron/native/wgc-capture/src/monitor_utils.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstdint>
 #include <windows.h>
 #include <string>
 #include <vector>
@@ -15,5 +14,5 @@ struct MonitorInfo {
 };
 
 std::vector<MonitorInfo> enumerateMonitors();
-HMONITOR findMonitorByDisplayId(int64_t displayId);
+HMONITOR findMonitorByDisplayId(int displayId);
 MonitorInfo getMonitorInfo(HMONITOR monitor);
